ml_service/requirements.txt
xgboost
pandas
pymongo
fastapi
uvicorn
scikit-learn
pydantic
python-dotenv
google-generativeai
